pyproject.toml
[project]
name = "alteza-proefopdracht"
version = "0.1.0"
description = "A Django application to fetch and display GitHub commits for a specified repository."
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0.3",
    "django-allauth[socialaccount]>=65.15.1",
    "django-environ>=0.13.0",
    "django-tailwind[cookiecutter,honcho,reload]>=4.4.2",
    "flake8>=7.3.0",
    "httpx>=0.28.1",
    "pygithub>=2.9.0",
    "pytest>=9.0.2",
    "pytest-django>=4.12.0",
    "pytest-xdist>=3.8.0",
    "ruff>=0.15.9",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "alteza_proefopdracht.settings.tests"
python_files = "tests.py test_*.py *_tests.py"
addopts = "-v --tb=short -n auto"

[dependency-groups]
dev = [
    "django-stubs[compatible-mypy]>=6.0.2",
    "djlint>=1.36.4",
]

[tool.djlint]
profile="django"
